+++ python/openassetio/__init__.py
@@ -73,6 +73,7 @@ The documentation for OpenAssetIO can be found here:
 # TODO(DF): @pylint
 from ._openassetio import TraitsData  # pylint: disable=import-error
 from ._openassetio import Context  # pylint: disable=import-error
+from ._openassetio import EntityReference  # pylint: disable=import-error
 
 # pylint: disable=wrong-import-position
 from .SpecificationBase import SpecificationBase