+++ tests/python/openassetio/test_entityreference.py
@@ -1,0 +1,41 @@
+#
+#   Copyright 2022 The Foundry Visionmongers Ltd
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+"""
+Tests for the EntityReference type.
+"""
+
+# pylint: disable=no-self-use, too-few-public-methods
+# pylint: disable=invalid-name,redefined-outer-name
+# pylint: disable=missing-class-docstring,missing-function-docstring
+
+import pytest
+
+from openassetio import EntityReference
+
+
+class Test_EntityReference_inheritance:
+    def test_class_is_final(self):
+        with pytest.raises(TypeError):
+
+            class _(EntityReference):
+                pass
+
+
+class Test_EntityReference_toString:
+    def test_returns_constructor_supplied_string(self):
+        expected = "ref://some/entity/🐈"
+        an_entity_reference = EntityReference(expected)
+        assert an_entity_reference.toString() == expected